+++ packages/button/src/DeleteButton.tsx
@@ -8,7 +8,7 @@
 import React, { ButtonHTMLAttributes } from 'react';
 import styled from '@emotion/styled';
 import { colors } from '@ndla/core';
-import { Button } from './Button';
+import { Button } from './OldButton';
 
 const StyledDeleteButton = styled(Button)`
   color: ${colors.support.red};
